// Hrms.Web/src/app/admin/asset/model/asset.model.ts
export class AssetModel {
    id: number;
    name: string;
    purchaseDate: string;
    isInWarranty: boolean;
    warrantyPeriod: number;
    serialNumber: string;
    assetTypeId: number;
    assetTypeName: string;
    variantId: number;
    variantName: string;
    manufacturerId: number;
    manufacturerName: string;
    status: number;
    vendorName: string;
    lastAssignDate: string;

}